feat(data,platform): ToTicks returns i64

+++ src/Cobalt.Engine/data/src/db/mod.rs
@@ -241,9 +241,9 @@ impl AlertManager {
         &mut self,
         times: &impl TimeSystem,
     ) -> Result<Vec<TriggeredAlert>> {
-        let day_start = times.day_start().to_ticks() as i64;
-        let week_start = times.week_start().to_ticks() as i64;
-        let month_start = times.month_start().to_ticks() as i64;
+        let day_start = times.day_start().to_ticks();
+        let week_start = times.week_start().to_ticks();
+        let month_start = times.month_start().to_ticks();
         let result = query_as(include_str!("../queries/triggered_alerts.sql"))
             .bind(day_start)
             .bind(week_start)
@@ -258,9 +258,9 @@ impl AlertManager {
         &mut self,
         times: &impl TimeSystem,
     ) -> Result<Vec<TriggeredReminder>> {
-        let day_start = times.day_start().to_ticks() as i64;
-        let week_start = times.week_start().to_ticks() as i64;
-        let month_start = times.month_start().to_ticks() as i64;
+        let day_start = times.day_start().to_ticks();
+        let week_start = times.week_start().to_ticks();
+        let month_start = times.month_start().to_ticks();
         let result = query_as(include_str!("../queries/triggered_reminders.sql"))
             .bind(day_start)
             .bind(week_start)
